Crate commonlibsse_ng

This is a Rust reimplementation of CommonLibSSE-NG.

It is intended to be memory-safe using the power of Rust.

rel: Module related to Relocation (calculate address from ID according to version, read module information, parse version information, etc.)

rex: Module related to Win32 API

skse: Module related to SKSE.exe (version information of SkyrimSE.exe, etc.)

sys: C++ binding (when no_sys feature is disabled): This is enabled by default. However, currently the types generated by rust-bindgen are difficult to use, and since the inline function does not exist in the .lib, only the inline function can be called.

Modules

rel

REL dir portion of CommonLibSSE-NG written by hand.

rex

Module related to Win32 API

skse

Module related to SKSE.exe (version information of SkyrimSE.exe, etc.)