commonlibsse_ng::re::ActorValues

Struct ActorValue_CEnum

#[repr(transparent)]
pub struct ActorValue_CEnum(pub u32);

Auto-generated FFI type for ActorValue.

When use this?

C’s enum is really just a number and there is no guarantee that it will fit within the enum. Therefore, it is used for the following cases that cannot be controlled by Rust.

C++ members

Function return values.

When not to use it?

Mainly those that can be controlled for safety on the Rust side.

Function Arguments

Convenient methods

to_enum/from_enum: To inter-convert enums.

count: Returns the number of defined discriminants.

Memory Layout

It will always have #[repr(transparent)]. In other words, it is equivalent to the size specified in repr.

Tuple Fields
0: u32
Implementations

impl ActorValue_CEnum

pub const None: Self

pub const Aggression: Self

pub const Confidence: Self

pub const Energy: Self

pub const Morality: Self

pub const Mood: Self

pub const Assistance: Self

pub const OneHanded: Self

pub const TwoHanded: Self

pub const Archery: Self

pub const Block: Self

pub const Smithing: Self

pub const HeavyArmor: Self

pub const LightArmor: Self

pub const Pickpocket: Self

pub const Lockpicking: Self

pub const Sneak: Self

pub const Alchemy: Self

pub const Speech: Self

pub const Alteration: Self

pub const Conjuration: Self

pub const Destruction: Self

pub const Illusion: Self

pub const Restoration: Self

pub const Enchanting: Self

pub const Health: Self

pub const Magicka: Self

pub const Stamina: Self

pub const HealRate: Self

pub const MagickaRate: Self

pub const StaminaRate: Self

pub const SpeedMult: Self

pub const InventoryWeight: Self

pub const CarryWeight: Self

pub const CriticalChance: Self

pub const MeleeDamage: Self

pub const UnarmedDamage: Self

pub const Mass: Self

pub const VoicePoints: Self

pub const VoiceRate: Self

pub const DamageResist: Self

pub const PoisonResist: Self

pub const ResistFire: Self

pub const ResistShock: Self

pub const ResistFrost: Self

pub const ResistMagic: Self

pub const ResistDisease: Self

pub const PerceptionCondition: Self

pub const EnduranceCondition: Self

pub const LeftAttackCondition: Self

pub const RightAttackCondition: Self

pub const LeftMobilityCondition: Self

pub const RightMobilityCondition: Self

pub const BrainCondition: Self

pub const Paralysis: Self

pub const Invisibility: Self

pub const NightEye: Self

pub const DetectLifeRange: Self

pub const WaterBreathing: Self

pub const WaterWalking: Self

pub const IgnoreCrippledLimbs: Self

pub const Fame: Self

pub const Infamy: Self

pub const JumpingBonus: Self

pub const WardPower: Self

pub const RightItemCharge: Self

pub const ArmorPerks: Self

pub const ShieldPerks: Self

pub const WardDeflection: Self

pub const Variable01: Self

pub const Variable02: Self

pub const Variable03: Self

pub const Variable04: Self

pub const Variable05: Self

pub const Variable06: Self

pub const Variable07: Self

pub const Variable08: Self

pub const Variable09: Self

pub const Variable10: Self

pub const BowSpeedBonus: Self

pub const FavorActive: Self

pub const FavorsPerDay: Self

pub const FavorsPerDayTimer: Self

pub const LeftItemCharge: Self

pub const AbsorbChance: Self

pub const Blindness: Self

pub const WeaponSpeedMult: Self

pub const ShoutRecoveryMult: Self

pub const BowStaggerBonus: Self

pub const Telekinesis: Self

pub const FavorPointsBonus: Self

pub const LastBribedIntimidated: Self

pub const LastFlattered: Self

pub const MovementNoiseMult: Self

pub const BypassVendorStolenCheck: Self

pub const BypassVendorKeywordCheck: Self

pub const WaitingForPlayer: Self

pub const OneHandedModifier: Self

pub const TwoHandedModifier: Self

pub const MarksmanModifier: Self

pub const BlockModifier: Self

pub const SmithingModifier: Self

pub const HeavyArmorModifier: Self

pub const LightArmorModifier: Self

pub const PickpocketModifier: Self

pub const LockpickingModifier: Self

pub const SneakingModifier: Self

pub const AlchemyModifier: Self

pub const SpeechcraftModifier: Self

pub const AlterationModifier: Self

pub const ConjurationModifier: Self

pub const DestructionModifier: Self

pub const IllusionModifier: Self

pub const RestorationModifier: Self

pub const EnchantingModifier: Self

pub const OneHandedSkillAdvance: Self

pub const TwoHandedSkillAdvance: Self

pub const MarksmanSkillAdvance: Self

pub const BlockSkillAdvance: Self

pub const SmithingSkillAdvance: Self

pub const HeavyArmorSkillAdvance: Self

pub const LightArmorSkillAdvance: Self

pub const PickpocketSkillAdvance: Self

pub const LockpickingSkillAdvance: Self

pub const SneakingSkillAdvance: Self

pub const AlchemySkillAdvance: Self

pub const SpeechcraftSkillAdvance: Self

pub const AlterationSkillAdvance: Self

pub const ConjurationSkillAdvance: Self

pub const DestructionSkillAdvance: Self

pub const IllusionSkillAdvance: Self

pub const RestorationSkillAdvance: Self

pub const EnchantingSkillAdvance: Self

pub const LeftWeaponSpeedMultiply: Self

pub const DragonSouls: Self

pub const CombatHealthRegenMultiply: Self

pub const OneHandedPowerModifier: Self

pub const TwoHandedPowerModifier: Self

pub const MarksmanPowerModifier: Self

pub const BlockPowerModifier: Self

pub const SmithingPowerModifier: Self

pub const HeavyArmorPowerModifier: Self

pub const LightArmorPowerModifier: Self

pub const PickpocketPowerModifier: Self

pub const LockpickingPowerModifier: Self

pub const SneakingPowerModifier: Self

pub const AlchemyPowerModifier: Self

pub const SpeechcraftPowerModifier: Self

pub const AlterationPowerModifier: Self

pub const ConjurationPowerModifier: Self

pub const DestructionPowerModifier: Self

pub const IllusionPowerModifier: Self

pub const RestorationPowerModifier: Self

pub const EnchantingPowerModifier: Self

pub const DragonRend: Self

pub const AttackDamageMult: Self

pub const HealRateMult: Self

pub const MagickaRateMult: Self

pub const StaminaRateMult: Self

pub const WerewolfPerks: Self

pub const VampirePerks: Self

pub const GrabActorOffset: Self

pub const Grabbed: Self

pub const DEPRECATED05: Self

pub const ReflectDamage: Self

pub const fn to_enum(self) -> Option<ActorValue>

Converts to the corresponding enum variant.

Returns Some(ActorValue) if the value is valid, otherwise None.

pub const fn from_enum(e: ActorValue) -> Self

Creates the struct from the enum.

This allows for easy conversion back to the FFI-friendly representation.

pub const fn count() -> usize

Number of discriminant in enum.

Returns 165