commonlibsse_ng::re::BGSEntryPoint

Struct ENTRY_POINT_CEnum

#[repr(transparent)]
pub struct ENTRY_POINT_CEnum(pub u32);

Auto-generated FFI type for ENTRY_POINT.

When use this?

C’s enum is really just a number and there is no guarantee that it will fit within the enum. Therefore, it is used for the following cases that cannot be controlled by Rust.

C++ members

Function return values.

When not to use it?

Mainly those that can be controlled for safety on the Rust side.

Function Arguments

Convenient methods

to_enum/from_enum: To inter-convert enums.

count: Returns the number of defined discriminants.

Memory Layout

It will always have #[repr(transparent)]. In other words, it is equivalent to the size specified in repr.

Tuple Fields
0: u32
Implementations

impl ENTRY_POINT_CEnum

pub const CalculateWeaponDamage: Self

pub const CalculateMyCriticalHitChance: Self

pub const CalculateMyCriticalHitDamage: Self

pub const CalculateMineExplodeChance: Self

pub const AdjustLimbDamage: Self

pub const AdjustBookSkillPoints: Self

pub const ModRecoveredHealth: Self

pub const GetShouldAttack: Self

pub const ModBuyPrices: Self

pub const AddLeveledListOnDeath: Self

pub const GetMaxCarryWeight: Self

pub const ModAddictionChance: Self

pub const ModAddictionDuration: Self

pub const ModPositiveChemDuration: Self

pub const Activate: Self

pub const IgnoreRunningDuringDetection: Self

pub const IgnoreBrokenLock: Self

pub const ModEnemyCriticalHitChance: Self

pub const ModSneakAttackMult: Self

pub const ModMaxPlaceableMines: Self

pub const ModBowZoom: Self

pub const ModRecoverArrowChance: Self

pub const ModSkillUse: Self

pub const ModTelekinesisDistance: Self

pub const ModTelekinesisDamageMult: Self

pub const ModTelekinesisDamage: Self

pub const ModBashingDamage: Self

pub const ModPowerAttackStamina: Self

pub const ModPowerAttackDamage: Self

pub const ModSpellMagnitude: Self

pub const ModSpellDuration: Self

pub const ModSecondaryValueWeight: Self

pub const ModArmorWeight: Self

pub const ModIncomingStagger: Self

pub const ModTargetStagger: Self

pub const ModAttackDamage: Self

pub const ModIncomingDamage: Self

pub const ModTargetDamageResistance: Self

pub const ModSpellCost: Self

pub const ModPercentBlocked: Self

pub const ModShieldDeflectArrowChance: Self

pub const ModIncomingSpellMagnitude: Self

pub const ModIncomingSpellDuration: Self

pub const ModPlayerIntimidation: Self

pub const ModPlayerReputation: Self

pub const ModFavorPoints: Self

pub const ModBribeAmount: Self

pub const ModDetectionLight: Self

pub const ModDetectionMovement: Self

pub const ModSoulGemRecharge: Self

pub const SetSweepAttack: Self

pub const ApplyCombatHitSpell: Self

pub const ApplyBashingSpell: Self

pub const ApplyReanimateSpell: Self

pub const SetBooleanGraphVariable: Self

pub const ModSpellCastingSoundEvent: Self

pub const ModPickpocketChance: Self

pub const ModDetectionSneakSkill: Self

pub const ModFallingDamage: Self

pub const ModLockpickSweetSpot: Self

pub const ModSellPrices: Self

pub const CanPickpocketEquippedItem: Self

pub const ModLockpickLevelAllowed: Self

pub const SetLockpickStartingArc: Self

pub const SetProgressionPicking: Self

pub const MakeLockpicksUnbreakable: Self

pub const ModAlchemyEffectiveness: Self

pub const ApplyWeaponSwingSpell: Self

pub const ModCommandedActorLimit: Self

pub const ApplySneakingSpell: Self

pub const ModPlayerMagicSlowdown: Self

pub const ModWardMagickaAbsorptionPct: Self

pub const ModInitialIngredientEffectsLearned: Self

pub const PurifyAlchemyIngredients: Self

pub const FilterActivation: Self

pub const CanDualCastSpell: Self

pub const ModTemperingHealth: Self

pub const ModEnchantmentPower: Self

pub const ModSoulPctCapturedToWeapon: Self

pub const ModSoulGemEnchanting: Self

pub const ModNumberAppliedEnchantmentsAllowed: Self

pub const SetActivateLabel: Self

pub const ModShoutOK: Self

pub const ModPoisonDoseCount: Self

pub const ShouldApplyPlacedItem: Self

pub const ModArmorRating: Self

pub const ModLockpickingCrimeChance: Self

pub const ModIngredientsHarvested: Self

pub const ModSpellRange_TargetLoc: Self

pub const ModPotionsCreated: Self

pub const ModLockpickingKeyRewardChance: Self

pub const AllowMountActor: Self

pub const fn to_enum(self) -> Option<ENTRY_POINT>

Converts to the corresponding enum variant.

Returns Some(ENTRY_POINT) if the value is valid, otherwise None.

pub const fn from_enum(e: ENTRY_POINT) -> Self

Creates the struct from the enum.

This allows for easy conversion back to the FFI-friendly representation.

pub const fn count() -> usize

Number of discriminant in enum.

Returns 92