commonlibsse_ng::re::BSFixedString

Type Alias BSFixedString

pub type BSFixedString = BSFixedStringInternal<U8>;

A fixed-length C string.

This type is string-interlaced and reference-shares memory with the same string.

Encoding

Since this is an FFI type, the encoding is not guaranteed. It may be UTF-8, ANSI, or some other platform-specific encoding, as it uses char internally.

Therefore, when using this string, it is recommended to convert it to &str to handle it safely.

ref: Converting esp/string Encoding

Aliased Type

struct BSFixedString { /* private fields */ }

Implementations

impl BSFixedString

pub const DEFAULT: Self

const default

pub fn new(data: &CStr) -> Self

Creates a new BSFixedString from a &CStr.

Memory allocation

If the same string exists inside the allocator, no new memory allocation occurs.

pub unsafe fn new_unchecked(data: *const c_char) -> Self

Creates a new BSFixedString from a raw pointer.

Safety

data must be a valid null-terminated char string.

Memory allocation

If the same string exists inside the allocator, no new memory allocation occurs.

pub fn as_c_str(&self) -> &CStr

Gets the string as a CStr.

Encoding

The returned string may not be UTF-8 due to its FFI nature. It is safe to use this as a raw CStr, but converting to str should be done carefully.

pub fn to_str(&self) -> Option<&str>

Converts the string to &str if it is valid UTF-8.

pub fn contains(&self, rhs: &CStr) -> bool

Returns true if CStr passed as argument is contained in this string or not.