commonlibsse_ng::re::BSIntrusiveRefCounted

Trait BSIntrusiveRefCountedTrait

pub trait BSIntrusiveRefCountedTrait {
    // Required methods
    fn inc_ref(&self) -> u32;
    fn dec_ref(&self) -> u32;
}

Required Methods

fn inc_ref(&self) -> u32

Returns the value after +1 to the current value.

Note

Implementations must ensure that the returned value is the new reference count after incrementing, not the old one.

fn dec_ref(&self) -> u32

Decrements the reference count and returns the value after decrementing.

Note

Implementations must ensure that the returned value is the new reference count after decrementing, not the old one.