commonlibsse_ng::re::BSPointerHandle

Struct BSPointerHandle

#[repr(C)]
pub struct BSPointerHandle<T> { /* private fields */ }

A typed handle referencing an object of type T.

This provides type safety for handles, and can be used to retrieve a smart pointer or reference.

Implementations

impl<T> BSPointerHandle<T>

pub const fn new() -> Self

Creates a new, null handle.

pub const fn from_handle(handle: BSUntypedPointerHandle) -> Self

Creates a handle from an untyped handle.

pub const fn native_handle(&self) -> u32

Returns the raw native handle value.

pub const fn has_value(&self) -> bool

Returns true if the handle has a non-zero value.

pub const fn reset(&mut self)

Resets the handle to 0.

impl<T: RefCountable> BSPointerHandle<T>

pub fn get(&self) -> NiPointer<T>

Retrieves a smart pointer to the underlying object.

pub fn as_ref(&self) -> Option<&T>

Attempts to get a shared reference to the underlying object.