commonlibsse_ng::re::BSTCaseInsensitiveStringMap

Trait BSTCaseInsensitiveStringMapTrait

pub trait BSTCaseInsensitiveStringMapTrait<T> {
    // Required methods
    fn hash_function(&self, key: &str) -> u32;
    fn key_eq(&self, lhs: &str, rhs: &str) -> bool;
}

Trait to represent map operations.

Required Methods

fn hash_function(&self, key: &str) -> u32

Computes the hash of the key in a case-insensitive manner.

Arguments

key: The string key.

Returns

The hash value.

fn key_eq(&self, lhs: &str, rhs: &str) -> bool

Checks if two keys are equal in a case-insensitive manner.

Arguments

lhs: The left-hand side key.

rhs: The right-hand side key.

Returns

true if equal, false otherwise.