commonlibsse_ng::re::BSTSmartPointer

Struct BSTSmartPointer

#[repr(C)]
pub struct BSTSmartPointer<T, M = BSTSmartPointerIntrusiveRefCount>
where
    T: BSIntrusiveRefCountedTrait,
    M: ManageBSTSmartPointer,
{ /* private fields */ }

Smart pointer with customizable reference management.

This smart pointer optionally manages reference counts depending on the M parameter.

Panics

Dereferencing a null pointer via Deref or DerefMut will panic. Use [as_ref] or [is_null] to avoid this panic.

Implementations

impl<T, M> BSTSmartPointer<T, M>
where T: BSIntrusiveRefCountedTrait, M: ManageBSTSmartPointer,

pub unsafe fn new(ptr: *mut T) -> Self

Creates a new smart pointer from a raw pointer.

Safety

The pointer must be created from TESBox(MemoryManager::allocate).

pub fn from_non_null(ptr: NonNull<T>) -> Self

Creates a new smart pointer from a raw pointer.

pub fn reset(&mut self)

Replaces the internal pointer with null and releases the old object.

After calling this, [is_null] returns true.

pub fn from_box(value: TESBox<T>) -> Self

Creates a smart pointer from a TESBox, taking ownership.

pub const fn as_ref(&self) -> Option<&T>

Returns an immutable reference to the managed object, or None if the pointer is null.

Use this to safely access the internal object without risking a panic.

pub const fn as_mut(&mut self) -> Option<&mut T>

Returns a mutable reference to the managed object, or None if the pointer is null.

pub const fn as_ptr(&self) -> *mut T

Returns the raw pointer to the managed object.

pub const fn is_null(&self) -> bool

Returns true if the internal pointer is null.