commonlibsse_ng::re::ControlMap

Struct ControlMap

#[repr(C)]
pub struct ControlMap {
    pub __base: [u8; 1],
    pub __base1: BSTEventSource<UserEventEnabled>,
    pub controlMap: [*mut InputContext; 17],
}

Fields
__base: [u8; 1]__base1: BSTEventSource<UserEventEnabled>controlMap: [*mut InputContext; 17]
Implementations

impl ControlMap

pub fn get_singleton() -> Option<&'static ControlMap>

Gets the singleton instance of ControlMap.

pub fn get_singleton_mut() -> Option<&'static mut ControlMap>

Gets the mutable singleton instance of ControlMap.

pub fn get_runtime_data(&self) -> Result<&RUNTIME_DATA, RelocationError>

Gets fields whose offset is determined at runtime.

Errors

This function may return an error if the module’s runtime is not available or if any error occurs while fetching the runtime state. Specifically, it calls ModuleState::map_active, which could result in an error.

pub fn get_runtime_data_mut( &mut self, ) -> Result<&mut RUNTIME_DATA, RelocationError>

Gets mutable fields whose offset is determined at runtime.

Errors

This function may return an error if the module’s runtime is not available or if any error occurs while fetching the runtime state. Specifically, it calls ModuleState::map_active_mut, which could result in an error.

pub fn allow_text_input(&mut self, allow: bool) -> Option<i8>

pub fn get_button_name_from_user_event( &self, event_id: &BSFixedString, device: INPUT_DEVICE, ) -> Option<BSFixedString>

pub fn get_user_event_name( &self, button_id: u32, device: INPUT_DEVICE, context: INPUT_CONTEXT_ID, ) -> Option<&BSFixedString>

pub fn pop_input_context(&mut self, context: INPUT_CONTEXT_ID) -> bool

pub fn push_input_context(&mut self, context: INPUT_CONTEXT_ID) -> bool