commonlibsse_ng::re::FormTypes

Struct FormType_CEnum

#[repr(transparent)]
pub struct FormType_CEnum(pub u8);

Auto-generated FFI type for FormType.

When use this?

C’s enum is really just a number and there is no guarantee that it will fit within the enum. Therefore, it is used for the following cases that cannot be controlled by Rust.

C++ members

Function return values.

When not to use it?

Mainly those that can be controlled for safety on the Rust side.

Function Arguments

Convenient methods

to_enum/from_enum: To inter-convert enums.

count: Returns the number of defined discriminants.

Memory Layout

It will always have #[repr(transparent)]. In other words, it is equivalent to the size specified in repr.

Tuple Fields
0: u8
Implementations

impl FormType_CEnum

pub const None: Self

pub const PluginInfo: Self

pub const FormGroup: Self

pub const GameSetting: Self

pub const Keyword: Self

pub const LocationRefType: Self

pub const Action: Self

pub const TextureSet: Self

pub const MenuIcon: Self

pub const Global: Self

pub const Class: Self

pub const Faction: Self

pub const HeadPart: Self

pub const Eyes: Self

pub const Race: Self

pub const Sound: Self

pub const AcousticSpace: Self

pub const Skill: Self

pub const MagicEffect: Self

pub const Script: Self

pub const LandTexture: Self

pub const Enchantment: Self

pub const Spell: Self

pub const Scroll: Self

pub const Activator: Self

pub const TalkingActivator: Self

pub const Armor: Self

pub const Book: Self

pub const Container: Self

pub const Door: Self

pub const Ingredient: Self

pub const Light: Self

pub const Misc: Self

pub const Apparatus: Self

pub const Static: Self

pub const StaticCollection: Self

pub const MovableStatic: Self

pub const Grass: Self

pub const Tree: Self

pub const Flora: Self

pub const Furniture: Self

pub const Weapon: Self

pub const Ammo: Self

pub const NPC: Self

pub const LeveledNPC: Self

pub const KeyMaster: Self

pub const AlchemyItem: Self

pub const IdleMarker: Self

pub const Note: Self

pub const ConstructibleObject: Self

pub const Projectile: Self

pub const Hazard: Self

pub const SoulGem: Self

pub const LeveledItem: Self

pub const Weather: Self

pub const Climate: Self

pub const ShaderParticleGeometryData: Self

pub const ReferenceEffect: Self

pub const Region: Self

pub const Navigation: Self

pub const Cell: Self

pub const Reference: Self

pub const ActorCharacter: Self

pub const ProjectileMissile: Self

pub const ProjectileArrow: Self

pub const ProjectileGrenade: Self

pub const ProjectileBeam: Self

pub const ProjectileFlame: Self

pub const ProjectileCone: Self

pub const ProjectileBarrier: Self

pub const PlacedHazard: Self

pub const WorldSpace: Self

pub const Land: Self

pub const NavMesh: Self

pub const TLOD: Self

pub const Dialogue: Self

pub const Info: Self

pub const Quest: Self

pub const Idle: Self

pub const Package: Self

pub const CombatStyle: Self

pub const LoadScreen: Self

pub const LeveledSpell: Self

pub const AnimatedObject: Self

pub const Water: Self

pub const EffectShader: Self

pub const TOFT: Self

pub const Explosion: Self

pub const Debris: Self

pub const ImageSpace: Self

pub const ImageAdapter: Self

pub const FormList: Self

pub const Perk: Self

pub const BodyPartData: Self

pub const AddonNode: Self

pub const ActorValueInfo: Self

pub const CameraShot: Self

pub const CameraPath: Self

pub const VoiceType: Self

pub const MaterialType: Self

pub const Impact: Self

pub const ImpactDataSet: Self

pub const Armature: Self

pub const EncounterZone: Self

pub const Location: Self

pub const Message: Self

pub const Ragdoll: Self

pub const DefaultObject: Self

pub const LightingMaster: Self

pub const MusicType: Self

pub const Footstep: Self

pub const FootstepSet: Self

pub const StoryManagerBranchNode: Self

pub const StoryManagerQuestNode: Self

pub const StoryManagerEventNode: Self

pub const DialogueBranch: Self

pub const MusicTrack: Self

pub const DialogueView: Self

pub const WordOfPower: Self

pub const Shout: Self

pub const EquipSlot: Self

pub const Relationship: Self

pub const Scene: Self

pub const AssociationType: Self

pub const Outfit: Self

pub const ArtObject: Self

pub const MaterialObject: Self

pub const MovementType: Self

pub const SoundRecord: Self

pub const DualCastData: Self

pub const SoundCategory: Self

pub const SoundOutputModel: Self

pub const CollisionLayer: Self

pub const ColorForm: Self

pub const ReverbParam: Self

pub const LensFlare: Self

pub const LensSprite: Self

pub const VolumetricLighting: Self

pub const Max: Self

pub const fn to_enum(self) -> Option<FormType>

Converts to the corresponding enum variant.

Returns Some(FormType) if the value is valid, otherwise None.

pub const fn from_enum(e: FormType) -> Self

Creates the struct from the enum.

This allows for easy conversion back to the FFI-friendly representation.

pub const fn count() -> usize

Number of discriminant in enum.

Returns 139