commonlibsse_ng::re::IAnimationGraphManagerHolder

Struct IAnimationGraphManagerHolderVtbl

#[repr(C)]
pub struct IAnimationGraphManagerHolderVtbl {

    pub CxxDrop: fn(this: &mut IAnimationGraphManagerHolder),
    pub NotifyAnimationGraph: fn(this: &IAnimationGraphManagerHolder, event_name: &BSFixedString) -> bool,
    pub GetAnimationGraphManagerImpl: fn(this: &IAnimationGraphManagerHolder, out: &mut BSTSmartPointer<BSAnimationGraphManager>) -> bool,
    pub SetAnimationGraphManagerImpl: fn(this: &mut IAnimationGraphManagerHolder, in_mgr: BSTSmartPointer<BSAnimationGraphManager>) -> bool,
    pub PopulateGraphNodesToTarget: fn(this: &IAnimationGraphManagerHolder, nodes: &mut BSScrapArray<*mut NiAVObject>) -> bool,
    pub ConstructAnimationGraph: fn(this: &IAnimationGraphManagerHolder, out: &mut BSTSmartPointer<BShkbAnimationGraph>) -> bool,
    pub Unk_06: fn(this: &IAnimationGraphManagerHolder),
    pub Unk_07: fn(this: &IAnimationGraphManagerHolder),
    pub SetupAnimEventSinks: fn(this: &IAnimationGraphManagerHolder, anim_graph: &BSTSmartPointer<BShkbAnimationGraph>) -> bool,
    pub Unk_09: fn(this: &IAnimationGraphManagerHolder),
    pub CreateAnimationChannels: fn(this: &IAnimationGraphManagerHolder, channels: &mut BSScrapArray<BSTSmartPointer<BSAnimationGraphChannel>>) -> bool,
    pub PostCreateAnimationGraphManager: fn(this: &IAnimationGraphManagerHolder, anim_graph_mgr: &mut BSTSmartPointer<BSAnimationGraphManager>),
    pub Unk_0C: fn(this: &IAnimationGraphManagerHolder),
    pub PostChangeAnimationManager: fn(this: &IAnimationGraphManagerHolder, arg1: &BSTSmartPointer<BShkbAnimationGraph>, arg2: &BSTSmartPointer<BShkbAnimationGraph>),
    pub Unk_0E: fn(this: &IAnimationGraphManagerHolder),
    pub GetGraphVariableCacheSize: fn(this: &IAnimationGraphManagerHolder) -> u32,
    pub GetGraphVariableImpl1: fn(this: &IAnimationGraphManagerHolder, name: &BSFixedString, out: &mut f32) -> bool,
    pub GetGraphVariableImpl2: fn(this: &IAnimationGraphManagerHolder, name: &BSFixedString, out: &mut i32) -> bool,
    pub GetGraphVariableImpl3: fn(this: &IAnimationGraphManagerHolder, name: &BSFixedString, out: &mut bool) -> bool,

}

The virtual function table for IAnimationGraphManagerHolder.

This struct defines function pointers to simulate the C++ virtual functions.

Fields
CxxDrop: fn(this: &mut IAnimationGraphManagerHolder)

Destructor.

NotifyAnimationGraph: fn(this: &IAnimationGraphManagerHolder, event_name: &BSFixedString) -> bool

Virtual functions corresponding to the C++ methods.

GetAnimationGraphManagerImpl: fn(this: &IAnimationGraphManagerHolder, out: &mut BSTSmartPointer<BSAnimationGraphManager>) -> boolSetAnimationGraphManagerImpl: fn(this: &mut IAnimationGraphManagerHolder, in_mgr: BSTSmartPointer<BSAnimationGraphManager>) -> boolPopulateGraphNodesToTarget: fn(this: &IAnimationGraphManagerHolder, nodes: &mut BSScrapArray<*mut NiAVObject>) -> boolConstructAnimationGraph: fn(this: &IAnimationGraphManagerHolder, out: &mut BSTSmartPointer<BShkbAnimationGraph>) -> boolUnk_06: fn(this: &IAnimationGraphManagerHolder)Unk_07: fn(this: &IAnimationGraphManagerHolder)SetupAnimEventSinks: fn(this: &IAnimationGraphManagerHolder, anim_graph: &BSTSmartPointer<BShkbAnimationGraph>) -> boolUnk_09: fn(this: &IAnimationGraphManagerHolder)CreateAnimationChannels: fn(this: &IAnimationGraphManagerHolder, channels: &mut BSScrapArray<BSTSmartPointer<BSAnimationGraphChannel>>) -> boolPostCreateAnimationGraphManager: fn(this: &IAnimationGraphManagerHolder, anim_graph_mgr: &mut BSTSmartPointer<BSAnimationGraphManager>)Unk_0C: fn(this: &IAnimationGraphManagerHolder)PostChangeAnimationManager: fn(this: &IAnimationGraphManagerHolder, arg1: &BSTSmartPointer<BShkbAnimationGraph>, arg2: &BSTSmartPointer<BShkbAnimationGraph>)Unk_0E: fn(this: &IAnimationGraphManagerHolder)GetGraphVariableCacheSize: fn(this: &IAnimationGraphManagerHolder) -> u32GetGraphVariableImpl1: fn(this: &IAnimationGraphManagerHolder, name: &BSFixedString, out: &mut f32) -> boolGetGraphVariableImpl2: fn(this: &IAnimationGraphManagerHolder, name: &BSFixedString, out: &mut i32) -> boolGetGraphVariableImpl3: fn(this: &IAnimationGraphManagerHolder, name: &BSFixedString, out: &mut bool) -> bool
Implementations

impl IAnimationGraphManagerHolderVtbl

pub const fn new() -> Self

Creates a new default virtual table with stubbed functions.