commonlibsse_ng::re::MemoryManager

Struct TESGlobalAlloc

pub struct TESGlobalAlloc;

The Skyrim global memory allocator using MemoryManager C++ class.

It implements Allocator and GlobalAlloc, so it can be used for #[global_allocator] and other Allocator changeable arrays.

Note: while this type is unstable, the functionality it provides can be accessed through the free functions in alloc.

CI

Skyrim MemoryManager is not available for CI, Therefore, the test_on_ci feature is enabled, it will automatically fall back to Rust’s Global.