commonlibsse_ng::re::NiAVObject

Struct NiAVObject

#[repr(C)]
pub struct NiAVObject {

    pub __base: NiObjectNET,
    pub parent: *mut NiNode,
    pub parent_index: u32,
    pub unk03C: u32,
    pub collision_object: NiPointer<NiCollisionObject>,
    pub local_transform: NiTransform,
    pub world_transform: NiTransform,
    pub previous_world_transform: NiTransform,
    pub world_bound: NiBound,
    pub unk_f4: u32,
    pub unk_f8: u64,
    pub fade_amount: c_float,
    pub last_updated_frame_counter: u32,
    pub unk104: u64,

}

Fields
__base: NiObjectNETparent: *mut NiNodeparent_index: u32unk03C: u32collision_object: NiPointer<NiCollisionObject>local_transform: NiTransformworld_transform: NiTransformprevious_world_transform: NiTransformworld_bound: NiBoundunk_f4: u32unk_f8: u64fade_amount: c_floatlast_updated_frame_counter: u32unk104: u64
Implementations

impl NiAVObject

pub const RTTI: VariantID = RTTI_NiObject

pub const NI_RTTI: VariantID = NiRTTI_NiObject

pub const VTABLE: [VariantID; 1] = VTABLE_NiObject

pub fn clone(&self) -> *mut NiAVObject

pub fn get_app_culled(&self) -> bool

Returns whether the object is culled (hidden) in the application.

This function checks the visibility status of the object based on its flags. If the object has the Flag::Hidden set, it is considered culled (i.e., hidden) and will not be rendered or processed for certain operations.

The get_app_culled function essentially checks if the object is flagged as “hidden” by returning a boolean indicating whether the object is culled or not. This is commonly used for optimizing rendering performance by excluding objects that should not be visible.

Returns

true if the object is culled (i.e., hidden).

false if the object is not culled (i.e., visible).

pub fn get_collision_object(&self) -> *mut bhkCollisionObject

pub fn get_flags(&self) -> Result<&Flag, RelocationError>

Errors

pub fn get_flags_mut(&mut self) -> Result<&mut Flag, RelocationError>

Errors