commonlibsse_ng::re::UI

Struct UI

#[repr(C)]
pub struct UI {

    pub __base: [u8; 8],
    pub __base1: BSTEventSource<MenuOpenCloseEvent>,
    pub __base2: BSTEventSource<MenuModeChangeEvent>,
    pub __base3: BSTEventSource<*mut c_void>,
    pub menuStack: BSTArray<GPtr<IMenu>>,
    pub menuMap: BSTHashMap<BSFixedString, UIMenuEntry>,
    pub processMessagesLock: BSSpinLock,
    pub numPausesGame: u32,
    pub numItemMenus: u32,
    pub numDisablePauseMenu: u32,
    pub numAllowSaving: u32,
    pub numDontHideCursorWhenTopmost: u32,
    pub numCustomRendering: u32,
    pub numApplicationMenus: u32,
    pub modal: bool,
    pub pad17D: u8,
    pub pad17E: u16,
    pub uiTimer: BSTimer,
    pub menuSystemVisible: bool,
    pub closingAllMenus: bool,
    pub pad1C2: u16,
    pub pad1C4: u32,
    pub unk1C8: u32,
    pub unk1CA: u32,

}

Fields
__base: [u8; 8]__base1: BSTEventSource<MenuOpenCloseEvent>__base2: BSTEventSource<MenuModeChangeEvent>__base3: BSTEventSource<*mut c_void>menuStack: BSTArray<GPtr<IMenu>>menuMap: BSTHashMap<BSFixedString, UIMenuEntry>processMessagesLock: BSSpinLocknumPausesGame: u32numItemMenus: u32numDisablePauseMenu: u32numAllowSaving: u32numDontHideCursorWhenTopmost: u32numCustomRendering: u32numApplicationMenus: u32modal: boolpad17D: u8pad17E: u16uiTimer: BSTimermenuSystemVisible: boolclosingAllMenus: boolpad1C2: u16pad1C4: u32unk1C8: u32unk1CA: u32
Implementations

impl UI

pub fn get_singleton() -> Option<&'static UI>

Returns a reference to the singleton instance of UI, if available.

pub fn get_singleton_mut() -> Option<&'static mut UI>

Returns a mutable reference to the singleton instance of UI, if available.

pub const fn game_is_paused(&self) -> bool

Returns true if the game is currently paused.

pub fn get_menu(&self, menu_name: &CStr) -> Option<GPtr<IMenu>>

Gets a pointer to the specified menu if it is currently registered.

pub fn get_movie_view(&self, menu_name: &CStr) -> Option<GPtr<GFxMovieView>>

Gets the GFxMovieView associated with the specified menu.

pub const fn is_cursor_hidden_when_topmost(&self) -> bool

Returns true if the cursor should be hidden when this UI is the topmost.

pub const fn is_item_menu_open(&self) -> bool

Returns true if an inventory/item menu is currently open.

pub fn is_menu_open(&self, menu_name: &CStr) -> bool

Checks whether a given menu is currently open and on the stack.

pub const fn is_modal_menu_open(&self) -> bool

Returns true if any modal menu is currently open.

pub const fn is_pause_menu_disabled(&self) -> bool

Returns true if the pause menu is currently disabled.

pub const fn is_saving_allowed(&self) -> bool

Returns true if saving the game is currently allowed.

pub const fn is_showing_menus(&self) -> bool

Returns true if the menu system is currently being displayed.

pub const fn is_using_custom_rendering(&self) -> bool

Returns true if custom rendering is being used for menus.

pub fn register(&mut self, menu_name: &CStr, creator: fn() -> *mut IMenu)

Registers a new menu with the UI system.

pub const fn show_menus(&mut self, show: bool)

Sets whether the UI system should be visible.

pub fn get_menu_by_name<T>(&self, name: &CStr) -> Option<GPtr<T>>
where T: RefCounted,

pub fn get_menu_as<T>(&self) -> Option<GPtr<T>>
where T: HasMenuName + RefCounted,

impl UI

pub fn add_menu_open_close_sink( &mut self, sink: *mut BSTEventSink<MenuOpenCloseEvent>, )

pub fn remove_menu_open_close_sink( &mut self, sink: *mut BSTEventSink<MenuOpenCloseEvent>, )

pub fn add_menu_mode_change_sink( &mut self, sink: *mut BSTEventSink<MenuModeChangeEvent>, )

pub fn remove_menu_mode_change_sink( &mut self, sink: *mut BSTEventSink<MenuModeChangeEvent>, )

pub fn add_void_sink(&mut self, sink: *mut BSTEventSink<*mut c_void>)

pub fn remove_void_sink(&mut self, sink: *mut BSTEventSink<*mut c_void>)