commonlibsse_ng::re

Module hkArray

hkArray

This module defines the hkArray struct, a dynamic array-like structure that mimics the behavior of a C++ container with methods for array manipulation, including dynamic resizing, accessors, and iterators.

Structs

hkArray

A dynamic array container C++’s hkArray, backed by a custom allocator.

hkArrayBase

Represents a base class for a dynamic array of type T.

hkArrayDrain

Iterator returned by hkArray::drain()

hkArrayIterMut

hkArrayIterator

hkArrayRefIterator