commonlibsse_ng::rel::id
Module shared_rwlock

Structs

MappedRwLockReadGuard

RAII structure used to release the shared read access of a lock when dropped, which can point to a subfield of the protected data.

MappedRwLockWriteGuard

RAII structure used to release the exclusive write access of a lock when dropped, which can point to a subfield of the protected data.

PoisonError

A type of error which can be returned whenever a lock is acquired.

RwLockReadGuard

RAII structure used to release the shared read access of a lock when dropped.

RwLockWriteGuard

RAII structure used to release the exclusive write access of a lock when dropped.

SharedRwLock

It exists in the SharedMemory situation and atomically edits the data involved in the lock.

Enums

MemoryMapError

Defines errors that may occur when working with MemoryMap.

TryLockError

An enumeration of possible errors associated with a TryLockResult which can occur while trying to acquire a lock, from the try_lock method on a Mutex or the try_read and try_write methods on an RwLock.

Type Aliases

LockResult

A type alias for the result of a lock method which can be poisoned.

TryLockResult

A type alias for the result of a nonblocking locking method.