commonlibsse_ng::rel::module
Struct ModuleHandle

pub struct ModuleHandle(/* private fields */);

Wrapper type to safely hold and handle valid handle addresses provided by GetModuleHandleW.

Unsafe this implementation

The module handle is the start of the exe, but if you donโt know the end, you donโt know how far is the valid memory range. The current implementation could crash at any time. It holds void ptr internally, but can be handled null-safely by getter by method.

Why not use HMODULE as it is?

It is not thread-safe as it is because it holds raw_pointer.

Therefore, we can keep it safe by creating another type that is valid as long as it holds the pointer, with the restriction that it is only invalidated on drop.

Implementations

impl ModuleHandle

pub fn new<H>(module_name: H) -> Result<Self, ModuleHandleError>
where H: Param<PCWSTR>,

Gets the module handle of a module (exe, dll, etc.) that is being loaded by the calling process.

Example

use commonlibsse_ng::rel::module::ModuleHandle;
use windows::core::h; // `h!` is utf-16 str macro.

let handle = ModuleHandle::new(h!("kernel32.dll"));
assert!(handle.is_ok());

// If there is no extension, a `.dll` is automatically specified.(This is the behavior of `GetModuleHandleW` function.)
let handle = ModuleHandle::new(h!("kernel32"));
assert!(handle.is_ok());

Errors

Errors if a module is specified that is not loaded by the calling process.

If the specified module handle could not be obtained.

pub const fn to_hmodule(&self) -> HMODULE

Returns the raw HMODULE handle.

pub const fn as_raw(&self) -> usize

Returns the module handle itself (i.e., the virtual address of the exe located in the DRAM).

pub fn try_as_nt_header(&self) -> Result<&IMAGE_NT_HEADERS64, ModuleHandleError>

Attempt to parse NT Header part.

Errors

When fail to parse as valid header.