commonlibsse_ng::rel

Module relocation

Structs

PhantomMember

A zero-sized marker used to access dynamically relocated members.

Relocation

Enums

RawPointerError

Represents errors that may occur when converting raw pointers to references.

RelocationError

Represents errors that may occur during member relocation.

Constants

INT3

JMP8

JMP32

NOP

NOP2

NOP3

NOP4

NOP5

NOP6

NOP7

NOP8

NOP9

RET

Functions

raw_pointer_as_mut⚠

Converts a raw pointer to a mutable reference.

raw_pointer_as_ref⚠

Converts a raw pointer to an immutable reference.

relocate

Errors

relocate_member

Relocates a member based on the runtime state, returning a pointer to the new location.

relocate_member_if⚠

Relocates a member based on a condition, using either offset a or b depending on the condition.

relocate_member_if_newer⚠

Relocates a member based on the version comparison, using either older or newer offset depending on the current version.

relocate_member_if_newer_mut⚠

Relocates a member based on the version comparison, using either older or newer offset depending on the current version.

relocate_member_mut

Relocates a member based on the runtime state, returning a pointer to the new location.

relocate_virtual⚠

Safety