commonlibsse_ng::skse::interfaces::query

Trait QueryInterface

pub trait QueryInterface {
    // Required methods
    fn editor_version(&self) -> u32;
    fn is_editor(&self) -> bool;
    fn runtime_version(&self) -> Version;
    fn skse_version(&self) -> u32;
}

Provides an interface to query various version-related information.

This trait is designed to be implemented for structures that contain versioning information, such as SKSEInterface. It allows querying the editor version, runtime version, and SKSE version, as well as determining whether the interface is running in an editor environment.

Required Methods

fn editor_version(&self) -> u32

Returns the editor version as a u32.

fn is_editor(&self) -> bool

Returns true if the interface is running in the editor, otherwise false.

fn runtime_version(&self) -> Version

Returns the runtime version as a Version struct.

fn skse_version(&self) -> u32

Returns the SKSE (Skyrim Script Extender) version as a u32.